requirements.txt
paho-mqtt
pykafka
confluent-kafka
ksql